added screen shake slider
Once upon a time in 1789 a man named Lorents Magnus Schultz was born. Upon discovery of a boat, he decided to sail around the entire world. And found the country of China. And when he stepped on the ground *BAM* a massive earthquake appeared and he was shaking, but the came up with the idea of shaking the screen in video games when video games were created, unfortunetly he died, and his ideas never came to be, THE END.